requirements.txt
# Django and Web Framework
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
asgiref==3.8.1

# Database - Using compatible versions
djongo==1.3.6
pymongo==4.5.0

# IMPORTANT: Use sqlparse 0.2.4 for djongo compatibility
sqlparse==0.2.4

# Cache and Message Broker
redis==5.0.1

# Trading and Data
ccxt==4.2.85
pandas==2.2.0
numpy==1.26.3

# Machine Learning
scikit-learn==1.5.0
joblib==1.4.2

# Authentication
djangorestframework-simplejwt==5.3.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3